@@ -7,13 +7,13 @@
     "upgrade-dependents": "./bin/upgrade-dependents.js"
   },
   "scripts": {
-    "unit": "tap -J -R=spec \"./test/**/*.spec.js\"",
+    "unit": "TS_NODE_FILES=true tap -J -R=spec \"./test/**/*.spec.js\"",
     "test": "npm run unit -- --cov",
     "lint": "tslint --project .",
     "clean": "rimraf ./lib",
     "build": "tsc",
     "report": "nyc report --reporter=lcov",
-    "ci": "npm run lint && npm run build && npm run test && npm run report"
+    "ci": "npm run build && npm run test && npm run report"
   },
   "repository": {
     "type": "git",
@@ -39,33 +39,34 @@
   },
   "homepage": "https://github.com/morrisallison/upgrade-dependents#readme",
   "engines": {
-    "node": ">=7.6.0"
+    "node": "^7.6.0"
   },
   "dependencies": {
-    "find-yarn-workspace-root": "^1.1.0",
-    "get-monorepo-packages": "^1.0.1",
+    "find-yarn-workspace-root": "^1.2.1",
+    "get-monorepo-packages": "^1.2.0",
     "json-file-plus": "^3.3.1",
-    "p-limit": "^2.0.0",
-    "read-pkg": "^3.0.0",
-    "semver": "^5.5.0",
-    "semver-utils": "^1.1.2",
-    "yargs": "^11.0.0"
+    "p-limit": "^2.3.0",
+    "read-pkg": "^5.2.0",
+    "semver": "^7.3.2",
+    "semver-utils": "^1.1.4",
+    "yargs": "^15.3.1"
   },
   "devDependencies": {
-    "@types/node": "^10.3.4",
-    "@types/p-limit": "^2.0.0",
-    "@types/read-pkg": "^3.0.0",
-    "@types/semver": "^5.5.0",
-    "@types/yargs": "^11.0.0",
-    "expect": "^23.1.0",
+    "@types/json-file-plus": "^3.3.2",
+    "@types/node": "^14.0.9",
+    "@types/p-limit": "^2.2.0",
+    "@types/read-pkg": "^5.1.0",
+    "@types/semver": "^7.2.0",
+    "@types/yargs": "^15.0.5",
+    "expect": "^26.0.1",
     "prettier": "1.13.6",
-    "rimraf": "^2.6.2",
-    "tap": "^12.0.1",
-    "testdouble": "^3.8.1",
-    "ts-node": "^6.2.0",
-    "tslint": "^5.10.0",
-    "tslint-config-prettier": "^1.13.0",
-    "typescript": "^2.9.2"
+    "rimraf": "^3.0.2",
+    "tap": "^14.10.7",
+    "testdouble": "^3.16.0",
+    "ts-node": "^8.10.2",
+    "tslint": "^6.1.2",
+    "tslint-config-prettier": "^1.18.0",
+    "typescript": "^3.9.3"
   },
   "files": [
     "bin",
