@@ -1,4 +1,6 @@
+import { PackageMeta } from "get-monorepo-packages";
+
 export type PackageInfo = {
   location: string;
-  meta: GetMonorepoPackages.PackageMeta;
+  meta: PackageMeta;
 };
